requirements.txt
cycler==0.11.0
fonttools==4.28.5
imageio==2.13.5
kiwisolver==1.3.2
matplotlib==3.5.1
networkx==2.6.3
numpy==1.21.5
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
PyWavelets==1.2.0
scikit-image==0.19.1
scipy==1.7.3
six==1.16.0
tifffile==2021.11.2
